+++ ui/src/app/modules/platform-tools/docker/container-restart/container-restart.component.ts
@@ -9,6 +9,7 @@ import { ToastrService } from 'ngx-toastr'
 
 @Component({
   templateUrl: './container-restart.component.html',
+  styleUrls: ['./container-restart.component.scss'],
 })
 export class ContainerRestartComponent implements OnInit, OnDestroy {
   checkTimeout: NodeJS.Timeout