+++ ui/src/app/modules/platform-tools/linux/linux.component.ts
@@ -1,8 +1,0 @@
-import { Component } from '@angular/core'
-
-@Component({
-  templateUrl: './linux.component.html',
-})
-export class LinuxComponent {
-  constructor() {}
-}