+++ CHANGELOG.md
@@ -9,6 +9,8 @@ All notable changes to `homebridge-config-ui-x` will be documented in this file.
 - fix login wallpaper setting field should in safari
   - it should not be offering to populate with a username/password from credentials manager
 - move docker startup script to settings
+- consistent designs for power options modules
+  - in particular, the `restart`, `docker-restart`, `linux-restart` and `linux-shutdown` modules
 
 ## v4.60.1 (2024-10-06)